some refactors and early eval

+++ src/movegen/bishop.rs
@@ -15,6 +15,6 @@ impl Mover for BishopType {
 
     #[inline]
     fn pseudo_legal_moves<CO: ColorType>(from: Square, pos: &Position) -> Bitboard {
-        get_bishop_moves(from, pos.occupancy) & !pos.by_color[CO::COLOR as usize]
+        get_bishop_moves(from, pos.occupancy) & !pos.by_color[CO::COLOR]
     }
 }
