+++ src/movegen/king.rs
@@ -15,7 +15,7 @@ impl Mover for KingType {
 
     #[inline]
     fn pseudo_legal_moves<CO: ColorType>(from: Square, pos: &Position) -> Bitboard {
-        get_king_moves(from) & !pos.by_color[CO::COLOR as usize]
+        get_king_moves(from) & !pos.by_color[CO::COLOR]
     }
 
     #[inline]
