+++ src/movegen/knight.rs
@@ -15,6 +15,6 @@ impl Mover for KnightType {
 
     #[inline]
     fn pseudo_legal_moves<CO: ColorType>(from: Square, pos: &Position) -> Bitboard {
-        get_knight_moves(from) & !pos.by_color[CO::COLOR as usize]
+        get_knight_moves(from) & !pos.by_color[CO::COLOR]
     }
 }
